Part 1
Most of the rewrite is done. Now it's just a matter of fixing the masks (make them work in position:absolute instead of fixed). I think in the end I'm gonna just go through all my changes that I made in the rewrite and merge them into the original code for my PR, that way I don't make to many changes to the repo (potentially breaking something)